Merge my branch

+++ src/main/java/com/example/demo/Ressource/FilmRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FilmRepository extends JpaRepository<Film, Integer> {
+public interface FilmRepository extends JpaRepository<Film, Long> {
 }
