+++ src/main/java/com/example/demo/Ressource/RealisateurRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RealisateurRepository extends JpaRepository<Realisateur, Integer> {
+public interface RealisateurRepository extends JpaRepository<Realisateur, Long> {
 }
